Divide the oversized modules in four more crates (#693)
Continues the crate-by-crate physical pass of #689 and #691. Six files were
over the 1,000-line review budget, each a single subject mixed with its
neighbours:

| file | lines | submodules | largest |
| --- | --- | --- | --- |
| wow-entities/unit_subsystems/movement.rs | 1,676 | subsystem, delayed_actions, generators, actions | 878 |
| wow-movement/spline.rs | 1,393 | move_spline, init, interpolation, fall_and_jump | 692 |
| wow-entities/item.rs | 1,247 | item, values, slots, templates | 906 |
| wow-entities/vehicle.rs | 1,076 | vehicle, seats, immunity, plans | 655 |
| bnet-server/rest/handlers.rs | 1,058 | login, tickets, wrong_password, srp, encoding, responses | 452 |
| wow-config/lib.rs | 1,053 | world_validation, world_registry, store | 444 |

Nothing outside the curated hotspot aggregates is over the budget in these
four crates now, and the physical ratchet's cohesion reviews drop from 82
to 76.

`wow-config/lib.rs` is a crate root, so its families are sibling modules
beside it and the root keeps its own file and `pub use` surface; a submodule
of a crate root imports `crate::*` rather than `super::*`.

Item-for-item movement under the visibility rules #689 and #691 established.
Three mechanics this pass needed:

- The item scanner had to learn `async fn`, `const fn` and `default` before
  the keyword. Nine top-level `async fn` handlers in the REST module were
  invisible to it, so they would have travelled with a neighbouring item
  instead of the family they belong to. The compiler found it as an
  unresolved call, not the scan.
- A `const` or `static` whose value continues on the following lines ends at
  the line that closes with `;`, not at its first line. The world-config
  registry's `include_str!` value proved that.
- `include_str!` and `include_bytes!` paths are re-resolved against the
  filesystem for each new file's depth rather than rewritten by pattern.

Fields, methods and one private type that a sibling family or a test root
reads are widened to `pub(super)`, the reach they had inside the single file,
each from the compiler's own E0616/E0451/E0624 diagnostic. Bindings and
parent globs the compiler then reported as unused are removed, so no target
gains a warning.

Item multiset identical per divided file (30, 48, 91, 52, 48 and 51 items).
Crate tests unchanged - wow-entities 725, wow-movement 128, bnet-server 87,
wow-config 39 - with no test path renamed. Architecture check, self-test,
physical ratchet, hotspot ratchet and the syntax-only session ownership
check pass.

Closes #693
