+++ crates/wow-entities/src/item/item.rs
@@ -1,337 +1,8 @@
-use std::collections::HashSet;
+//! Item packets.
+//!
+//! Separated from item.rs under #693.
 
-use wow_constants::{
-    BagFamilyMask, EnchantmentSlot, InventoryResult, InventoryType, ItemBondingType, ItemClass,
-    ItemContext, ItemFieldFlags, ItemFieldFlags2, ItemFlags, ItemModifier, ItemSubClassContainer,
-    ItemSubClassQuiver, ItemUpdateState, TypeId, TypeMask,
-};
-use wow_core::ObjectGuid;
-
-use crate::{
-    EntityObject, ObjectDataUpdate, UpdateMask,
-    update_fields::{ITEM_DATA_BITS, TYPEID_ITEM},
-};
-
-pub const MAX_ITEM_SPELLS: usize = 5;
-pub const MAX_ENCHANTMENT_SLOT: usize = 13;
-pub const MAX_INSPECTED_ENCHANTMENT_SLOT: usize = 8;
-pub const MAX_SPECIALIZATIONS: usize = 5;
-pub const ITEM_MODIFIER_COUNT: usize = 58;
-pub const BOP_TRADEABLE_DURATION_SECS: u32 = 2 * 60 * 60;
-pub const INVENTORY_SLOT_BAG_0: u8 = 255;
-pub const EQUIPMENT_SLOT_HEAD: u8 = 0;
-pub const EQUIPMENT_SLOT_NECK: u8 = 1;
-pub const EQUIPMENT_SLOT_SHOULDERS: u8 = 2;
-pub const EQUIPMENT_SLOT_BODY: u8 = 3;
-pub const EQUIPMENT_SLOT_CHEST: u8 = 4;
-pub const EQUIPMENT_SLOT_WAIST: u8 = 5;
-pub const EQUIPMENT_SLOT_LEGS: u8 = 6;
-pub const EQUIPMENT_SLOT_FEET: u8 = 7;
-pub const EQUIPMENT_SLOT_WRISTS: u8 = 8;
-pub const EQUIPMENT_SLOT_HANDS: u8 = 9;
-pub const EQUIPMENT_SLOT_FINGER1: u8 = 10;
-pub const EQUIPMENT_SLOT_FINGER2: u8 = 11;
-pub const EQUIPMENT_SLOT_TRINKET1: u8 = 12;
-pub const EQUIPMENT_SLOT_TRINKET2: u8 = 13;
-pub const EQUIPMENT_SLOT_BACK: u8 = 14;
-pub const EQUIPMENT_SLOT_MAINHAND: u8 = 15;
-pub const EQUIPMENT_SLOT_OFFHAND: u8 = 16;
-pub const EQUIPMENT_SLOT_RANGED: u8 = 17;
-pub const EQUIPMENT_SLOT_TABARD: u8 = 18;
-pub const EQUIPMENT_SLOT_END: u8 = 19;
-pub const PROFESSION_SLOT_PROFESSION1_TOOL: u8 = 19;
-pub const PROFESSION_SLOT_PROFESSION1_GEAR1: u8 = 20;
-pub const PROFESSION_SLOT_PROFESSION1_GEAR2: u8 = 21;
-pub const PROFESSION_SLOT_PROFESSION2_TOOL: u8 = 22;
-pub const PROFESSION_SLOT_PROFESSION2_GEAR1: u8 = 23;
-pub const PROFESSION_SLOT_PROFESSION2_GEAR2: u8 = 24;
-pub const PROFESSION_SLOT_COOKING_TOOL: u8 = 25;
-pub const PROFESSION_SLOT_COOKING_GEAR1: u8 = 26;
-pub const PROFESSION_SLOT_FISHING_TOOL: u8 = 27;
-pub const PROFESSION_SLOT_FISHING_GEAR1: u8 = 28;
-pub const PROFESSION_SLOT_FISHING_GEAR2: u8 = 29;
-pub const PROFESSION_SLOT_START: u8 = 19;
-pub const PROFESSION_SLOT_END: u8 = 30;
-pub const PROFESSION_SLOT_MAX_COUNT: u8 = 3;
-
-pub const ITEM_DATA_PARENT_BIT: usize = 0;
-pub const ITEM_DATA_ARTIFACT_POWERS_BIT: usize = 1;
-pub const ITEM_DATA_GEMS_BIT: usize = 2;
-pub const ITEM_DATA_OWNER_BIT: usize = 3;
-pub const ITEM_DATA_CONTAINED_IN_BIT: usize = 4;
-pub const ITEM_DATA_CREATOR_BIT: usize = 5;
-pub const ITEM_DATA_GIFT_CREATOR_BIT: usize = 6;
-pub const ITEM_DATA_STACK_COUNT_BIT: usize = 7;
-pub const ITEM_DATA_EXPIRATION_BIT: usize = 8;
-pub const ITEM_DATA_DYNAMIC_FLAGS_BIT: usize = 9;
-pub const ITEM_DATA_PROPERTY_SEED_BIT: usize = 10;
-pub const ITEM_DATA_RANDOM_PROPERTIES_ID_BIT: usize = 11;
-pub const ITEM_DATA_DURABILITY_BIT: usize = 12;
-pub const ITEM_DATA_MAX_DURABILITY_BIT: usize = 13;
-pub const ITEM_DATA_CREATE_PLAYED_TIME_BIT: usize = 14;
-pub const ITEM_DATA_CONTEXT_BIT: usize = 15;
-pub const ITEM_DATA_CREATE_TIME_BIT: usize = 16;
-pub const ITEM_DATA_ARTIFACT_XP_BIT: usize = 17;
-pub const ITEM_DATA_ITEM_APPEARANCE_MOD_ID_BIT: usize = 18;
-pub const ITEM_DATA_MODIFIERS_BIT: usize = 19;
-pub const ITEM_DATA_DYNAMIC_FLAGS2_BIT: usize = 20;
-pub const ITEM_DATA_ITEM_BONUS_KEY_BIT: usize = 21;
-pub const ITEM_DATA_DEBUG_ITEM_LEVEL_BIT: usize = 22;
-pub const ITEM_DATA_SPELL_CHARGES_PARENT_BIT: usize = 23;
-pub const ITEM_DATA_SPELL_CHARGES_FIRST_BIT: usize = 24;
-pub const ITEM_DATA_ENCHANTMENT_PARENT_BIT: usize = 29;
-pub const ITEM_DATA_ENCHANTMENT_FIRST_BIT: usize = 30;
-
-pub const ITEM_DATA_BASE_ALLOWED_MASK: [u32; 2] = [0xE029_CE7F, 0x0000_07FF];
-pub const ITEM_DATA_OWNER_ALLOWED_MASK: [u32; 2] = [0x1FD6_3180, 0x0000_0000];
-
-pub const APPEARANCE_MODIFIER_SLOT_BY_SPEC: [ItemModifier; MAX_SPECIALIZATIONS] = [
-    ItemModifier::TransmogAppearanceSpec1,
-    ItemModifier::TransmogAppearanceSpec2,
-    ItemModifier::TransmogAppearanceSpec3,
-    ItemModifier::TransmogAppearanceSpec4,
-    ItemModifier::TransmogAppearanceSpec5,
-];
-
-pub const ILLUSION_MODIFIER_SLOT_BY_SPEC: [ItemModifier; MAX_SPECIALIZATIONS] = [
-    ItemModifier::EnchantIllusionSpec1,
-    ItemModifier::EnchantIllusionSpec2,
-    ItemModifier::EnchantIllusionSpec3,
-    ItemModifier::EnchantIllusionSpec4,
-    ItemModifier::EnchantIllusionSpec5,
-];
-
-pub const SECONDARY_APPEARANCE_MODIFIER_SLOT_BY_SPEC: [ItemModifier; MAX_SPECIALIZATIONS] = [
-    ItemModifier::TransmogSecondaryAppearanceSpec1,
-    ItemModifier::TransmogSecondaryAppearanceSpec2,
-    ItemModifier::TransmogSecondaryAppearanceSpec3,
-    ItemModifier::TransmogSecondaryAppearanceSpec4,
-    ItemModifier::TransmogSecondaryAppearanceSpec5,
-];
-
-#[derive(Debug, Clone, PartialEq, Eq, Default)]
-pub struct ItemBonusKey {
-    pub item_id: i32,
-    pub bonus_list_ids: Vec<i32>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct ArtifactPower {
-    pub artifact_power_id: i16,
-    pub purchased_rank: u8,
-    pub current_rank_with_bonus: u8,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Default)]
-pub struct SocketedGem {
-    pub item_id: i32,
-    pub context: u8,
-    pub bonus_list_ids: Vec<u16>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-pub struct ItemEnchantment {
-    pub id: i32,
-    pub duration: u32,
-    pub charges: i16,
-    pub field_a: u8,
-    pub field_b: u8,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ItemDataValues {
-    pub artifact_powers: Vec<ArtifactPower>,
-    pub gems: Vec<SocketedGem>,
-    pub owner: ObjectGuid,
-    pub contained_in: ObjectGuid,
-    pub creator: ObjectGuid,
-    pub gift_creator: ObjectGuid,
-    pub stack_count: u32,
-    pub expiration: u32,
-    pub dynamic_flags: u32,
-    pub property_seed: i32,
-    pub random_properties_id: i32,
-    pub durability: u32,
-    pub max_durability: u32,
-    pub create_played_time: u32,
-    pub context: i32,
-    pub create_time: i64,
-    pub artifact_xp: u64,
-    pub item_appearance_mod_id: u8,
-    pub dynamic_flags2: u32,
-    pub modifiers: [u32; ITEM_MODIFIER_COUNT],
-    pub item_bonus_key: ItemBonusKey,
-    pub debug_item_level: u16,
-    pub spell_charges: [i32; MAX_ITEM_SPELLS],
-    pub enchantments: [ItemEnchantment; MAX_ENCHANTMENT_SLOT],
-}
-
-impl Default for ItemDataValues {
-    fn default() -> Self {
-        Self {
-            artifact_powers: Vec::new(),
-            gems: Vec::new(),
-            owner: ObjectGuid::EMPTY,
-            contained_in: ObjectGuid::EMPTY,
-            creator: ObjectGuid::EMPTY,
-            gift_creator: ObjectGuid::EMPTY,
-            stack_count: 0,
-            expiration: 0,
-            dynamic_flags: 0,
-            property_seed: 0,
-            random_properties_id: 0,
-            durability: 0,
-            max_durability: 0,
-            create_played_time: 0,
-            context: ItemContext::None as i32,
-            create_time: 0,
-            artifact_xp: 0,
-            item_appearance_mod_id: 0,
-            dynamic_flags2: 0,
-            modifiers: [0; ITEM_MODIFIER_COUNT],
-            item_bonus_key: ItemBonusKey::default(),
-            debug_item_level: 0,
-            spell_charges: [0; MAX_ITEM_SPELLS],
-            enchantments: [ItemEnchantment::default(); MAX_ENCHANTMENT_SLOT],
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ItemCreateInfo {
-    pub guid: ObjectGuid,
-    pub item_id: u32,
-    pub context: ItemContext,
-    pub owner: Option<ObjectGuid>,
-    pub max_durability: u32,
-    pub expiration: u32,
-    pub spell_charges: [i32; MAX_ITEM_SPELLS],
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct ItemStorageTemplate {
-    pub entry: u32,
-    pub class_id: ItemClass,
-    pub subclass_id: u32,
-    pub inventory_type: InventoryType,
-    pub bonding: ItemBondingType,
-    pub bag_family: BagFamilyMask,
-    pub max_stack_size: u32,
-    pub max_count: i32,
-    pub item_limit_category: u32,
-    pub container_slots: u8,
-    pub sell_price: u32,
-    pub is_crafting_reagent: bool,
-    pub flags: ItemFlags,
-}
-
-impl ItemStorageTemplate {
-    pub const fn regular_item(entry: u32, max_stack_size: u32) -> Self {
-        Self {
-            entry,
-            class_id: ItemClass::Miscellaneous,
-            subclass_id: 0,
-            inventory_type: InventoryType::NonEquip,
-            bonding: ItemBondingType::None,
-            bag_family: BagFamilyMask::NONE,
-            max_stack_size,
-            max_count: 0,
-            item_limit_category: 0,
-            container_slots: 0,
-            sell_price: 0,
-            is_crafting_reagent: false,
-            flags: ItemFlags::empty(),
-        }
-    }
-
-    pub fn is_bound_account_wide(&self) -> bool {
-        self.flags.contains(ItemFlags::IS_BOUND_TO_ACCOUNT)
-    }
-
-    pub const fn can_change_equip_state_in_combat(&self) -> bool {
-        matches!(
-            self.inventory_type,
-            InventoryType::Relic | InventoryType::Shield | InventoryType::Holdable
-        ) || matches!(self.class_id, ItemClass::Weapon | ItemClass::Projectile)
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct ItemDataUpdate {
-    pub mask: UpdateMask,
-    pub values: ItemDataValues,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct ItemValuesUpdate {
-    pub changed_object_type_mask: u32,
-    pub object_data: Option<ObjectDataUpdate>,
-    pub item_data: Option<ItemDataUpdate>,
-}
-
-impl ItemValuesUpdate {
-    pub const fn has_data(&self) -> bool {
-        self.changed_object_type_mask != 0
-    }
-}
-
-pub fn item_can_go_into_bag(proto: &ItemStorageTemplate, bag_proto: &ItemStorageTemplate) -> bool {
-    match bag_proto.class_id {
-        ItemClass::Container => match bag_proto.subclass_id {
-            subclass if subclass == ItemSubClassContainer::Container as u32 => true,
-            subclass if subclass == ItemSubClassContainer::SoulContainer as u32 => {
-                proto.bag_family.contains(BagFamilyMask::SOUL_SHARDS)
-            }
-            subclass if subclass == ItemSubClassContainer::HerbContainer as u32 => {
-                proto.bag_family.contains(BagFamilyMask::HERBS)
-            }
-            subclass if subclass == ItemSubClassContainer::EnchantingContainer as u32 => {
-                proto.bag_family.contains(BagFamilyMask::ENCHANTING_SUPP)
-            }
-            subclass if subclass == ItemSubClassContainer::MiningContainer as u32 => {
-                proto.bag_family.contains(BagFamilyMask::MINING_SUPP)
-            }
-            subclass if subclass == ItemSubClassContainer::EngineeringContainer as u32 => {
-                proto.bag_family.contains(BagFamilyMask::ENGINEERING_SUPP)
-            }
-            subclass if subclass == ItemSubClassContainer::GemContainer as u32 => {
-                proto.bag_family.contains(BagFamilyMask::GEMS)
-            }
-            subclass if subclass == ItemSubClassContainer::LeatherworkingContainer as u32 => proto
-                .bag_family
-                .contains(BagFamilyMask::LEATHERWORKING_SUPP),
-            subclass if subclass == ItemSubClassContainer::InscriptionContainer as u32 => {
-                proto.bag_family.contains(BagFamilyMask::INSCRIPTION_SUPP)
-            }
-            subclass if subclass == ItemSubClassContainer::TackleContainer as u32 => {
-                proto.bag_family.contains(BagFamilyMask::FISHING_SUPP)
-            }
-            subclass if subclass == ItemSubClassContainer::CookingContainer as u32 => {
-                proto.bag_family.contains(BagFamilyMask::COOKING_SUPP)
-            }
-            subclass if subclass == ItemSubClassContainer::ReagentContainer as u32 => {
-                proto.is_crafting_reagent
-            }
-            _ => false,
-        },
-        ItemClass::Quiver => match bag_proto.subclass_id {
-            subclass if subclass == ItemSubClassQuiver::Quiver as u32 => {
-                proto.bag_family.contains(BagFamilyMask::ARROWS)
-            }
-            subclass if subclass == ItemSubClassQuiver::AmmoPouch as u32 => {
-                proto.bag_family.contains(BagFamilyMask::BULLETS)
-            }
-            _ => false,
-        },
-        _ => false,
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum ItemStateTransition {
-    Updated,
-    PretendNeverExisted,
-}
+use super::*;
 
 #[derive(Debug, Clone, PartialEq)]
 pub struct Item {
@@ -1233,14 +904,3 @@ impl Item {
         item_modified_appearance_id
     }
 }
-
-fn spec_modifier(
-    active_talent_group: usize,
-    slots: &[ItemModifier; MAX_SPECIALIZATIONS],
-) -> ItemModifier {
-    slots.get(active_talent_group).copied().unwrap_or(slots[0])
-}
-
-#[cfg(test)]
-#[path = "item/tests/mod.rs"]
-mod tests;
