+++ crates/wow-entities/src/unit_subsystems/movement/mod.rs
@@ -1,0 +1,18 @@
+// Copyright (c) 2026 alseif0x
+// RustyCore — WoW WotLK 3.4.3 server in Rust
+// Based on TrinityCore protocol research (https://github.com/TrinityCore/TrinityCore)
+// Licensed under GPL v3 — https://www.gnu.org/licenses/gpl-3.0.html
+
+//! Unit motion master and movement generators.
+
+use super::*;
+
+mod actions;
+mod delayed_actions;
+mod generators;
+mod subsystem;
+
+pub use actions::*;
+pub use delayed_actions::*;
+pub use generators::*;
+pub use subsystem::*;
