+++ crates/wow-entities/src/vehicle/vehicle.rs
@@ -1,10 +1,8 @@
-use std::collections::BTreeMap;
+//! Vehicle packets.
+//!
+//! Separated from vehicle.rs under #693.
 
-use wow_constants::{MovementFlag2, TypeId};
-pub use wow_constants::{VehicleExitParameter, VehicleFlag};
-use wow_core::{ObjectGuid, Position};
-
-pub const MAX_VEHICLE_SEATS: usize = 8;
+use super::*;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum VehicleStatus {
@@ -38,386 +36,9 @@ impl PassengerInfo {
     }
 }
 
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub struct VehicleSeatAddon {
-    pub seat_orientation_offset: f32,
-    pub exit_parameter_x: f32,
-    pub exit_parameter_y: f32,
-    pub exit_parameter_z: f32,
-    pub exit_parameter_o: f32,
-    pub exit_parameter: VehicleExitParameter,
-}
-
-impl Default for VehicleSeatAddon {
-    fn default() -> Self {
-        Self {
-            seat_orientation_offset: 0.0,
-            exit_parameter_x: 0.0,
-            exit_parameter_y: 0.0,
-            exit_parameter_z: 0.0,
-            exit_parameter_o: 0.0,
-            exit_parameter: VehicleExitParameter::None,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub struct VehicleSeatInfo {
-    pub id: u32,
-    pub attachment_offset: Position,
-    pub can_enter_or_exit: bool,
-    pub usable_by_override: bool,
-    pub can_control: bool,
-    pub can_switch_from_seat: bool,
-    pub ejectable: bool,
-    pub disables_gravity: bool,
-    pub passenger_not_selectable: bool,
-    pub keep_pet: bool,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub struct VehicleSeat {
-    pub seat_info: VehicleSeatInfo,
-    pub seat_addon: VehicleSeatAddon,
-    pub passenger: PassengerInfo,
-}
-
-impl VehicleSeat {
-    pub const fn new(seat_info: VehicleSeatInfo, seat_addon: VehicleSeatAddon) -> Self {
-        Self {
-            seat_info,
-            seat_addon,
-            passenger: PassengerInfo::empty(),
-        }
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.passenger.is_empty()
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct VehicleAccessory {
-    pub accessory_entry: u32,
-    pub is_minion: bool,
-    pub summon_time_ms: u32,
-    pub seat_id: i8,
-    pub summoned_type: u8,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct VehicleAccessoryInstallPlan {
-    pub remove_all_passengers: bool,
-    pub accessories: Vec<VehicleAccessory>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct VehicleAccessorySummonPlan {
-    pub accessory: VehicleAccessory,
-    pub add_accessory_unit_mask: bool,
-    pub handle_spell_click_seat_id: i8,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct VehiclePendingJoinAbort {
-    pub passenger: ObjectGuid,
-    pub seat_id: i8,
-    pub target_vehicle_available: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Default)]
-pub struct VehiclePendingEventRemovalPlan {
-    pub scheduled_aborts: Vec<VehiclePendingJoinAbort>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct VehicleRemoveAllPassengersPlan {
-    pub pending_join_aborts: Vec<VehiclePendingJoinAbort>,
-    pub remove_control_vehicle_auras: bool,
-    pub forced_exit_passengers: Vec<ObjectGuid>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct VehiclePassengerAddPlan {
-    pub accepted: bool,
-    pub seat_id: Option<i8>,
-    pub scheduled_abort: bool,
-    pub displaced_passenger: Option<ObjectGuid>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum VehiclePassengerTransportReset {
-    None,
-    Reset,
-    InheritBaseTransport,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct VehiclePassengerRemovePlan {
-    pub seat_id: i8,
-    pub set_vehicle_none: bool,
-    pub restore_gravity: bool,
-    pub restore_interactible: bool,
-    pub restore_npc_flag: bool,
-    pub remove_charm: bool,
-    pub transport_reset: VehiclePassengerTransportReset,
-    pub cast_parachute: bool,
-    pub call_ai_passenger_boarded: bool,
-    pub call_on_remove_passenger_script: bool,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub struct VehiclePassengerRelocation {
-    pub passenger: ObjectGuid,
-    pub position: Position,
-    pub set_home_position: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct VehicleJoinAbortPlan {
-    pub remove_pending_event: bool,
-    pub remove_control_vehicle_aura: bool,
-    pub despawn_accessory: bool,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct VehicleJoinExecutePlan {
-    pub passenger: ObjectGuid,
-    pub seat_id: i8,
-    pub abort: Option<VehicleJoinAbortPlan>,
-    pub pending_seat_aborts: VehiclePendingEventRemovalPlan,
-    pub pending_passenger_aborts: VehiclePendingEventRemovalPlan,
-    pub exit_existing_vehicle: bool,
-    pub passenger_info: Option<PassengerInfo>,
-    pub remove_npc_flag: bool,
-    pub interrupt_generic_spell: bool,
-    pub interrupt_autorepeat_spell: bool,
-    pub remove_mount_interrupt_auras: bool,
-    pub remove_mounted_auras: bool,
-    pub player_drop_battleground_flag: bool,
-    pub player_stop_casting_charm: bool,
-    pub player_stop_casting_bind_sight: bool,
-    pub player_cancel_expected_vehicle_ride_aura: bool,
-    pub player_unsummon_temporary_pet: bool,
-    pub set_disable_gravity: bool,
-    pub transport_position: Option<Position>,
-    pub set_vehicle_charm: bool,
-    pub send_clear_target: bool,
-    pub set_root_controlled: bool,
-    pub launch_transport_enter_spline: bool,
-    pub transfer_threat_to_vehicle: bool,
-    pub call_ai_passenger_boarded: bool,
-    pub call_on_add_passenger_script: bool,
-    pub call_on_install_accessory_script: bool,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum VehicleSpellImmunityKind {
-    Effect,
-    State,
-    Mechanic,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct VehicleSpellImmunity {
-    pub kind: VehicleSpellImmunityKind,
-    pub spell_or_mechanic: i32,
-    pub apply: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Default)]
-pub struct VehicleImmunityPlan {
-    pub immunities: Vec<VehicleSpellImmunity>,
-    pub root: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct VehicleResetPlan {
-    pub immunity_plan: VehicleImmunityPlan,
-    pub accessory_install_plan: Option<VehicleAccessoryInstallPlan>,
-    pub call_on_reset_script: bool,
-}
-
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
 pub struct VehicleTemplate {
     pub despawn_delay_ms: i32,
-}
-
-pub fn vehicle_accessory_install_plan_like_cpp(
-    base_type_id: TypeId,
-    evading: bool,
-    accessories: &[VehicleAccessory],
-) -> VehicleAccessoryInstallPlan {
-    VehicleAccessoryInstallPlan {
-        remove_all_passengers: base_type_id == TypeId::Player || !evading,
-        accessories: accessories
-            .iter()
-            .copied()
-            .filter(|accessory| !evading || accessory.is_minion)
-            .collect(),
-    }
-}
-
-pub fn vehicle_base_movement_flags_like_cpp(vehicle_flags: u32) -> MovementFlag2 {
-    let mut movement_flags = MovementFlag2::empty();
-    if vehicle_flags & VehicleFlag::NoStrafe as u32 != 0 {
-        movement_flags |= MovementFlag2::NO_STRAFE;
-    }
-    if vehicle_flags & VehicleFlag::NoJumping as u32 != 0 {
-        movement_flags |= MovementFlag2::NO_JUMPING;
-    }
-    if vehicle_flags & VehicleFlag::FullSpeedTurning as u32 != 0 {
-        movement_flags |= MovementFlag2::FULL_SPEED_TURNING;
-    }
-    if vehicle_flags & VehicleFlag::AllowPitching as u32 != 0 {
-        movement_flags |= MovementFlag2::ALWAYS_ALLOW_PITCHING;
-    }
-    if vehicle_flags & VehicleFlag::FullSpeedPitching as u32 != 0 {
-        movement_flags |= MovementFlag2::FULL_SPEED_PITCHING;
-    }
-    movement_flags
-}
-
-pub const SPELL_EFFECT_HEAL_LIKE_CPP: i32 = 10;
-pub const SPELL_EFFECT_DISPEL_LIKE_CPP: i32 = 38;
-pub const SPELL_EFFECT_KNOCK_BACK_LIKE_CPP: i32 = 98;
-pub const SPELL_EFFECT_HEAL_PCT_LIKE_CPP: i32 = 136;
-pub const SPELL_EFFECT_KNOCK_BACK_DEST_LIKE_CPP: i32 = 144;
-pub const SPELL_AURA_PERIODIC_HEAL_LIKE_CPP: i32 = 8;
-pub const SPELL_AURA_DAMAGE_SHIELD_LIKE_CPP: i32 = 15;
-pub const SPELL_AURA_MOD_RESISTANCE_LIKE_CPP: i32 = 22;
-pub const SPELL_AURA_MOD_STAT_LIKE_CPP: i32 = 29;
-pub const SPELL_AURA_MOD_DECREASE_SPEED_LIKE_CPP: i32 = 33;
-pub const SPELL_AURA_SCHOOL_IMMUNITY_LIKE_CPP: i32 = 39;
-pub const SPELL_AURA_SCHOOL_ABSORB_LIKE_CPP: i32 = 69;
-pub const SPELL_AURA_SPLIT_DAMAGE_PCT_LIKE_CPP: i32 = 81;
-pub const SPELL_AURA_MOD_DAMAGE_PERCENT_TAKEN_LIKE_CPP: i32 = 87;
-pub const SPELL_AURA_MOD_UNATTACKABLE_LIKE_CPP: i32 = 93;
-pub const MECHANIC_BANISH_LIKE_CPP: i32 = 18;
-pub const MECHANIC_SHIELD_LIKE_CPP: i32 = 19;
-pub const MECHANIC_IMMUNE_SHIELD_LIKE_CPP: i32 = 29;
-
-pub fn vehicle_immunity_plan_like_cpp(
-    vehicle_id: u32,
-    is_mechanical_creature: bool,
-    is_world_boss: bool,
-) -> VehicleImmunityPlan {
-    use VehicleSpellImmunityKind::{Effect, Mechanic, State};
-
-    let mut plan = VehicleImmunityPlan::default();
-    plan.immunities.extend([
-        VehicleSpellImmunity {
-            kind: Effect,
-            spell_or_mechanic: SPELL_EFFECT_KNOCK_BACK_LIKE_CPP,
-            apply: true,
-        },
-        VehicleSpellImmunity {
-            kind: Effect,
-            spell_or_mechanic: SPELL_EFFECT_KNOCK_BACK_DEST_LIKE_CPP,
-            apply: true,
-        },
-    ]);
-
-    if is_mechanical_creature && !is_world_boss {
-        plan.immunities.extend([
-            VehicleSpellImmunity {
-                kind: Effect,
-                spell_or_mechanic: SPELL_EFFECT_HEAL_LIKE_CPP,
-                apply: true,
-            },
-            VehicleSpellImmunity {
-                kind: Effect,
-                spell_or_mechanic: SPELL_EFFECT_HEAL_PCT_LIKE_CPP,
-                apply: true,
-            },
-            VehicleSpellImmunity {
-                kind: Effect,
-                spell_or_mechanic: SPELL_EFFECT_DISPEL_LIKE_CPP,
-                apply: true,
-            },
-            VehicleSpellImmunity {
-                kind: State,
-                spell_or_mechanic: SPELL_AURA_PERIODIC_HEAL_LIKE_CPP,
-                apply: true,
-            },
-            VehicleSpellImmunity {
-                kind: State,
-                spell_or_mechanic: SPELL_AURA_SCHOOL_IMMUNITY_LIKE_CPP,
-                apply: true,
-            },
-            VehicleSpellImmunity {
-                kind: State,
-                spell_or_mechanic: SPELL_AURA_MOD_UNATTACKABLE_LIKE_CPP,
-                apply: true,
-            },
-            VehicleSpellImmunity {
-                kind: State,
-                spell_or_mechanic: SPELL_AURA_SCHOOL_ABSORB_LIKE_CPP,
-                apply: true,
-            },
-            VehicleSpellImmunity {
-                kind: Mechanic,
-                spell_or_mechanic: MECHANIC_BANISH_LIKE_CPP,
-                apply: true,
-            },
-            VehicleSpellImmunity {
-                kind: Mechanic,
-                spell_or_mechanic: MECHANIC_SHIELD_LIKE_CPP,
-                apply: true,
-            },
-            VehicleSpellImmunity {
-                kind: Mechanic,
-                spell_or_mechanic: MECHANIC_IMMUNE_SHIELD_LIKE_CPP,
-                apply: true,
-            },
-            VehicleSpellImmunity {
-                kind: State,
-                spell_or_mechanic: SPELL_AURA_DAMAGE_SHIELD_LIKE_CPP,
-                apply: true,
-            },
-            VehicleSpellImmunity {
-                kind: State,
-                spell_or_mechanic: SPELL_AURA_SPLIT_DAMAGE_PCT_LIKE_CPP,
-                apply: true,
-            },
-            VehicleSpellImmunity {
-                kind: State,
-                spell_or_mechanic: SPELL_AURA_MOD_RESISTANCE_LIKE_CPP,
-                apply: true,
-            },
-            VehicleSpellImmunity {
-                kind: State,
-                spell_or_mechanic: SPELL_AURA_MOD_STAT_LIKE_CPP,
-                apply: true,
-            },
-            VehicleSpellImmunity {
-                kind: State,
-                spell_or_mechanic: SPELL_AURA_MOD_DAMAGE_PERCENT_TAKEN_LIKE_CPP,
-                apply: true,
-            },
-        ]);
-    }
-
-    match vehicle_id {
-        160 | 244 | 510 | 452 | 543 => {
-            plan.root = true;
-            plan.immunities.push(VehicleSpellImmunity {
-                kind: State,
-                spell_or_mechanic: SPELL_AURA_MOD_DECREASE_SPEED_LIKE_CPP,
-                apply: true,
-            });
-        }
-        335 | 336 | 338 => {
-            plan.immunities.push(VehicleSpellImmunity {
-                kind: State,
-                spell_or_mechanic: SPELL_AURA_MOD_DAMAGE_PERCENT_TAKEN_LIKE_CPP,
-                apply: false,
-            });
-        }
-        _ => {}
-    }
-
-    plan
 }
 
 #[derive(Debug, Clone, PartialEq)]
@@ -427,9 +48,9 @@ pub struct Vehicle {
     base_position: Position,
     vehicle_id: u32,
     creature_entry: u32,
-    usable_seat_num: u32,
-    seats: BTreeMap<i8, VehicleSeat>,
-    status: VehicleStatus,
+    pub(super) usable_seat_num: u32,
+    pub(super) seats: BTreeMap<i8, VehicleSeat>,
+    pub(super) status: VehicleStatus,
     pending_join_events: BTreeMap<ObjectGuid, i8>,
 }
 
@@ -1032,44 +653,3 @@ impl Vehicle {
         calculate_passenger_offset(global, self.base_position)
     }
 }
-
-pub fn calculate_passenger_position(offset: Position, transport: Position) -> Position {
-    Position::new(
-        transport.x + offset.x * transport.orientation.cos()
-            - offset.y * transport.orientation.sin(),
-        transport.y
-            + offset.y * transport.orientation.cos()
-            + offset.x * transport.orientation.sin(),
-        transport.z + offset.z,
-        normalize_orientation(transport.orientation + offset.orientation),
-    )
-}
-
-pub fn calculate_passenger_offset(global: Position, transport: Position) -> Position {
-    let mut x = global.x - transport.x;
-    let mut y = global.y - transport.y;
-    let z = global.z - transport.z;
-    let orientation = normalize_orientation(global.orientation - transport.orientation);
-
-    let inx = x;
-    let iny = y;
-    let tan = transport.orientation.tan();
-    let denom = transport.orientation.cos() + transport.orientation.sin() * tan;
-    y = (iny - inx * tan) / denom;
-    x = (inx + iny * tan) / denom;
-
-    Position::new(x, y, z, orientation)
-}
-
-fn normalize_orientation(mut orientation: f32) -> f32 {
-    let tau = std::f32::consts::TAU;
-    orientation %= tau;
-    if orientation < 0.0 {
-        orientation += tau;
-    }
-    orientation
-}
-
-#[cfg(test)]
-#[path = "vehicle/tests/mod.rs"]
-mod tests;
