+++ crates/wow-movement/src/spline/move_spline.rs
@@ -1,16 +1,8 @@
-// C++ movement code stores durations/curve params as int32/float and casts between them.
-#![allow(clippy::cast_precision_loss, clippy::cast_possible_wrap)]
+//! Move spline packets.
+//!
+//! Separated from spline.rs under #693.
 
-use std::f32::consts::PI;
-
-use bitflags::bitflags;
-use wow_constants::movement::MovementFlag;
-use wow_core::{ObjectGuid, Position};
-
-pub const GRAVITY_LIKE_CPP: f32 = 19.291_105;
-const TERMINAL_VELOCITY_LIKE_CPP: f32 = 60.148_003;
-const TERMINAL_SAFE_FALL_VELOCITY_LIKE_CPP: f32 = 7.0;
-const MINIMAL_DURATION_MS_LIKE_CPP: i32 = 1;
+use super::*;
 
 bitflags! {
     #[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
@@ -111,486 +103,6 @@ impl MoveSplineFlag {
     }
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-#[repr(u8)]
-pub enum MonsterMoveType {
-    #[default]
-    Normal = 0,
-    FacingSpot = 1,
-    FacingTarget = 2,
-    FacingAngle = 3,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub struct FacingInfo {
-    pub spot: Position,
-    pub target: ObjectGuid,
-    pub angle: f32,
-    pub kind: MonsterMoveType,
-}
-
-impl Default for FacingInfo {
-    fn default() -> Self {
-        Self {
-            spot: Position::ZERO,
-            target: ObjectGuid::EMPTY,
-            angle: 0.0,
-            kind: MonsterMoveType::Normal,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-pub struct SpellEffectExtraData {
-    pub target: ObjectGuid,
-    pub spell_visual_id: u32,
-    pub progress_curve_id: u32,
-    pub parabolic_curve_id: u32,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-pub struct AnimTierTransition {
-    pub tier_transition_id: u32,
-    pub anim_tier: u8,
-}
-
-#[derive(Debug, Clone, PartialEq, Default)]
-pub struct MonsterMovePathData {
-    pub points: Vec<Position>,
-    pub packed_deltas: Vec<[f32; 3]>,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct MoveSplineInitArgs {
-    pub path: Vec<Position>,
-    pub facing: FacingInfo,
-    pub flags: MoveSplineFlag,
-    pub path_idx_offset: i32,
-    pub velocity: f32,
-    pub parabolic_amplitude: f32,
-    pub vertical_acceleration: f32,
-    pub effect_start_time_percent: f32,
-    pub effect_start_time_ms: i32,
-    pub spline_id: u32,
-    pub initial_orientation: f32,
-    pub spell_effect_extra: Option<SpellEffectExtraData>,
-    pub anim_tier: Option<AnimTierTransition>,
-    pub walk: bool,
-    pub has_velocity: bool,
-    pub transform_for_transport: bool,
-}
-
-impl Default for MoveSplineInitArgs {
-    fn default() -> Self {
-        Self {
-            path: Vec::with_capacity(16),
-            facing: FacingInfo::default(),
-            flags: MoveSplineFlag::SMOOTH_GROUND_PATH,
-            path_idx_offset: 0,
-            velocity: 0.0,
-            parabolic_amplitude: 0.0,
-            vertical_acceleration: 0.0,
-            effect_start_time_percent: 0.0,
-            effect_start_time_ms: 0,
-            spline_id: 0,
-            initial_orientation: 0.0,
-            spell_effect_extra: None,
-            anim_tier: None,
-            walk: false,
-            has_velocity: false,
-            transform_for_transport: true,
-        }
-    }
-}
-
-impl MoveSplineInitArgs {
-    #[must_use]
-    pub fn with_capacity(path_capacity: usize) -> Self {
-        Self {
-            path: Vec::with_capacity(path_capacity),
-            ..Self::default()
-        }
-    }
-
-    pub fn validate(&self) -> Result<(), MoveSplineValidationError> {
-        if self.path.len() <= 1 {
-            return Err(MoveSplineValidationError::PathTooShort);
-        }
-        if self.velocity < 0.01 {
-            return Err(MoveSplineValidationError::VelocityTooLow);
-        }
-        if !(0.0..=1.0).contains(&self.effect_start_time_percent) {
-            return Err(MoveSplineValidationError::EffectStartTimePercentOutOfRange);
-        }
-        self.check_path_lengths()?;
-        Ok(())
-    }
-
-    fn check_path_lengths(&self) -> Result<(), MoveSplineValidationError> {
-        if self.path.len() > 2 || self.facing.kind == MonsterMoveType::Normal {
-            for pair in self.path.windows(2) {
-                if distance_3d(pair[0], pair[1]) < 0.1 {
-                    return Err(MoveSplineValidationError::PathSegmentTooShort);
-                }
-            }
-        }
-        Ok(())
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub struct MoveSplineLaunchInput {
-    pub current_position: Position,
-    pub active_spline_position: Option<Position>,
-    pub movement_flags: MovementFlag,
-    pub selected_speed: f32,
-    pub run_speed: f32,
-    pub assistance_speed_factor: f32,
-    pub on_transport: bool,
-}
-
-impl MoveSplineLaunchInput {
-    #[must_use]
-    pub const fn new(current_position: Position) -> Self {
-        Self {
-            current_position,
-            active_spline_position: None,
-            movement_flags: MovementFlag::NONE,
-            selected_speed: 0.0,
-            run_speed: 0.0,
-            assistance_speed_factor: 1.0,
-            on_transport: false,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub struct MoveSplineLaunchResult {
-    pub real_position: Position,
-    pub movement_flags: MovementFlag,
-    pub duration_ms: i32,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub struct MoveSplineStopInput {
-    pub current_position: Position,
-    pub active_spline_position: Option<Position>,
-    pub on_transport: bool,
-}
-
-impl MoveSplineStopInput {
-    #[must_use]
-    pub const fn new(current_position: Position) -> Self {
-        Self {
-            current_position,
-            active_spline_position: None,
-            on_transport: false,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub struct MoveSplineStopResult {
-    pub position: Position,
-    pub spline_id: u32,
-    pub stop_distance_tolerance: u32,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub struct JumpSpeeds {
-    pub speed_xy: f32,
-    pub speed_z: f32,
-}
-
-pub fn compute_jump_max_height_like_cpp(speed_z: f32) -> f32 {
-    let move_time_half = speed_z / GRAVITY_LIKE_CPP;
-    -compute_fall_elevation(move_time_half, false, -speed_z)
-}
-
-pub fn calculate_jump_speeds_like_cpp(
-    distance: f32,
-    base_speed: f32,
-    current_speed: f32,
-    speed_multiplier: f32,
-    min_height: f32,
-    max_height: f32,
-) -> JumpSpeeds {
-    let speed_xy = (base_speed * 3.0 * speed_multiplier).min(28.0_f32.max(current_speed * 4.0));
-    let duration = distance / speed_xy;
-    let duration_sqr = duration * duration;
-    let height = if duration_sqr < min_height * 8.0 / GRAVITY_LIKE_CPP {
-        min_height
-    } else if duration_sqr > max_height * 8.0 / GRAVITY_LIKE_CPP {
-        max_height
-    } else {
-        GRAVITY_LIKE_CPP * duration_sqr / 8.0
-    };
-
-    JumpSpeeds {
-        speed_xy,
-        speed_z: (2.0 * GRAVITY_LIKE_CPP * height).sqrt(),
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum MoveSplineLaunchError {
-    EmptyPath,
-    Validation(MoveSplineValidationError),
-}
-
-impl From<MoveSplineValidationError> for MoveSplineLaunchError {
-    fn from(value: MoveSplineValidationError) -> Self {
-        Self::Validation(value)
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct MoveSplineInit {
-    pub args: MoveSplineInitArgs,
-}
-
-impl MoveSplineInit {
-    #[must_use]
-    pub fn new(spline_id: u32) -> Self {
-        Self {
-            args: MoveSplineInitArgs {
-                spline_id,
-                transform_for_transport: false,
-                flags: MoveSplineFlag::SMOOTH_GROUND_PATH,
-                ..MoveSplineInitArgs::default()
-            },
-        }
-    }
-
-    pub fn move_by_path<I>(&mut self, controls: I, path_offset: i32)
-    where
-        I: IntoIterator<Item = Position>,
-    {
-        self.args.path_idx_offset = path_offset;
-        self.args.path.clear();
-        self.args.path.extend(controls);
-    }
-
-    pub fn move_to(&mut self, destination: Position) {
-        self.args.path_idx_offset = 0;
-        self.args.path.resize(2, Position::new(0.0, 0.0, 0.0, 0.0));
-        self.args.path[1] = destination;
-    }
-
-    pub fn set_first_point_id(&mut self, point_id: i32) {
-        self.args.path_idx_offset = point_id;
-    }
-
-    pub fn set_velocity(&mut self, velocity: f32) {
-        self.args.velocity = velocity;
-        self.args.has_velocity = true;
-    }
-
-    pub fn set_walk(&mut self, enable: bool) {
-        self.args.walk = enable;
-    }
-
-    pub fn set_smooth(&mut self) {
-        self.args.flags.enable_catmull_rom();
-    }
-
-    pub fn set_uncompressed(&mut self) {
-        self.args.flags.insert(MoveSplineFlag::UNCOMPRESSED_PATH);
-    }
-
-    pub fn set_cyclic(&mut self) {
-        self.args.flags.insert(MoveSplineFlag::CYCLIC);
-    }
-
-    pub fn set_fly(&mut self) {
-        self.args.flags.enable_flying();
-    }
-
-    pub fn set_transport_enter(&mut self) {
-        self.args.flags.enable_transport_enter();
-    }
-
-    pub fn set_transport_exit(&mut self) {
-        self.args.flags.enable_transport_exit();
-    }
-
-    pub fn set_backward(&mut self) {
-        self.args.flags.insert(MoveSplineFlag::BACKWARD);
-    }
-
-    pub fn set_unlimited_speed(&mut self) {
-        self.args.flags.insert(MoveSplineFlag::UNLIMITED_SPEED);
-    }
-
-    pub fn set_orientation_fixed(&mut self, enable: bool) {
-        self.args
-            .flags
-            .set(MoveSplineFlag::ORIENTATION_FIXED, enable);
-    }
-
-    pub fn set_fall(&mut self, falling_slow: bool) {
-        self.args.flags.enable_falling();
-        self.args
-            .flags
-            .set(MoveSplineFlag::FALLING_SLOW, falling_slow);
-    }
-
-    pub fn set_parabolic(&mut self, amplitude: f32, time_shift: f32) {
-        self.args.effect_start_time_percent = time_shift;
-        self.args.parabolic_amplitude = amplitude;
-        self.args.vertical_acceleration = 0.0;
-        self.args.flags.enable_parabolic();
-    }
-
-    pub fn set_parabolic_vertical_acceleration(
-        &mut self,
-        vertical_acceleration: f32,
-        time_shift: f32,
-    ) {
-        self.args.effect_start_time_percent = time_shift;
-        self.args.parabolic_amplitude = 0.0;
-        self.args.vertical_acceleration = vertical_acceleration;
-        self.args.flags.enable_parabolic();
-    }
-
-    pub fn set_animation(
-        &mut self,
-        anim_tier: u8,
-        tier_transition_id: u32,
-        transition_start_time_ms: i32,
-    ) {
-        self.args.effect_start_time_percent = 0.0;
-        self.args.effect_start_time_ms = transition_start_time_ms;
-        self.args.anim_tier = Some(AnimTierTransition {
-            tier_transition_id,
-            anim_tier,
-        });
-        self.args.flags.enable_animation();
-    }
-
-    pub fn set_facing_spot(&mut self, spot: Position) {
-        self.args.facing.spot = spot;
-        self.args.facing.kind = MonsterMoveType::FacingSpot;
-    }
-
-    pub fn set_facing_target_with_angle(&mut self, target: ObjectGuid, absolute_angle: f32) {
-        self.args.facing.angle = absolute_angle;
-        self.args.facing.target = target;
-        self.args.facing.kind = MonsterMoveType::FacingTarget;
-    }
-
-    pub fn set_facing_angle(&mut self, angle: f32) {
-        self.args.facing.angle = wrap_angle_0_2pi(angle);
-        self.args.facing.kind = MonsterMoveType::FacingAngle;
-    }
-
-    pub fn set_spell_effect_extra_data(&mut self, spell_effect_extra: SpellEffectExtraData) {
-        self.args.spell_effect_extra = Some(spell_effect_extra);
-    }
-
-    pub fn disable_transport_path_transformations(&mut self) {
-        self.args.transform_for_transport = false;
-    }
-
-    pub fn launch(
-        &mut self,
-        move_spline: &mut MoveSpline,
-        input: MoveSplineLaunchInput,
-    ) -> Result<MoveSplineLaunchResult, MoveSplineLaunchError> {
-        let real_position = input
-            .active_spline_position
-            .unwrap_or(input.current_position);
-
-        if self.args.path.is_empty() {
-            return Err(MoveSplineLaunchError::EmptyPath);
-        }
-
-        self.args.path[0] = real_position;
-        self.args.initial_orientation = real_position.orientation;
-        self.args.flags.set(
-            MoveSplineFlag::ENTER_CYCLE,
-            self.args.flags.contains(MoveSplineFlag::CYCLIC),
-        );
-        move_spline.on_transport = input.on_transport;
-
-        let mut movement_flags = input.movement_flags;
-        if self.args.flags.contains(MoveSplineFlag::BACKWARD) {
-            movement_flags.remove(MovementFlag::FORWARD);
-            movement_flags.insert(MovementFlag::BACKWARD);
-        } else {
-            movement_flags.remove(MovementFlag::BACKWARD);
-            movement_flags.insert(MovementFlag::FORWARD);
-        }
-
-        if movement_flags.contains(MovementFlag::ROOT) {
-            movement_flags.remove(MovementFlag::MASK_MOVING);
-        }
-
-        if !self.args.has_velocity {
-            self.args.velocity = input.selected_speed * input.assistance_speed_factor;
-        }
-
-        self.args.velocity = self.args.velocity.min(self.speed_limit(input.run_speed));
-        move_spline.initialize(&self.args)?;
-
-        Ok(MoveSplineLaunchResult {
-            real_position,
-            movement_flags,
-            duration_ms: move_spline.duration_ms(),
-        })
-    }
-
-    pub fn stop(
-        &mut self,
-        move_spline: &mut MoveSpline,
-        input: MoveSplineStopInput,
-    ) -> Option<MoveSplineStopResult> {
-        if move_spline.finalized() {
-            return None;
-        }
-
-        let position = input
-            .active_spline_position
-            .unwrap_or(input.current_position);
-        self.args.flags = MoveSplineFlag::DONE;
-        move_spline.on_transport = input.on_transport;
-        if move_spline.initialize(&self.args).is_err() {
-            return None;
-        }
-
-        Some(MoveSplineStopResult {
-            position,
-            spline_id: move_spline.id(),
-            stop_distance_tolerance: 2,
-        })
-    }
-
-    fn speed_limit(&self, run_speed: f32) -> f32 {
-        if self.args.flags.contains(MoveSplineFlag::UNLIMITED_SPEED) {
-            return f32::MAX;
-        }
-
-        if self.args.flags.intersects(
-            MoveSplineFlag::FALLING
-                | MoveSplineFlag::CATMULLROM
-                | MoveSplineFlag::FLYING
-                | MoveSplineFlag::PARABOLIC,
-        ) {
-            return 50.0;
-        }
-
-        28.0_f32.max(run_speed * 4.0)
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum MoveSplineValidationError {
-    PathTooShort,
-    VelocityTooLow,
-    EffectStartTimePercentOutOfRange,
-    PathSegmentTooShort,
-}
-
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum SplineUpdateResult {
     None = 0x01,
@@ -600,13 +112,13 @@ pub enum SplineUpdateResult {
 }
 
 #[derive(Debug, Clone, PartialEq, Default)]
-struct SplineData {
-    points: Vec<Position>,
-    lengths: Vec<i32>,
-    first: i32,
-    last: i32,
-    cyclic: bool,
-    smooth: bool,
+pub(super) struct SplineData {
+    pub(super) points: Vec<Position>,
+    pub(super) lengths: Vec<i32>,
+    pub(super) first: i32,
+    pub(super) last: i32,
+    pub(super) cyclic: bool,
+    pub(super) smooth: bool,
 }
 
 impl SplineData {
@@ -626,7 +138,7 @@ impl SplineData {
         self.lengths[(index + 1) as usize] - self.lengths[index as usize]
     }
 
-    fn point(&self, index: i32) -> Position {
+    pub(super) fn point(&self, index: i32) -> Position {
         self.points[index as usize]
     }
 
@@ -662,7 +174,7 @@ impl SplineData {
 
 #[derive(Debug, Clone, PartialEq)]
 pub struct MoveSpline {
-    spline: SplineData,
+    pub(super) spline: SplineData,
     facing: FacingInfo,
     id: u32,
     flags: MoveSplineFlag,
@@ -1178,215 +690,3 @@ impl MoveSpline {
         }
     }
 }
-
-#[must_use]
-pub fn compute_fall_time(path_length: f32, is_safe_fall: bool) -> f32 {
-    if path_length < 0.0 {
-        return 0.0;
-    }
-
-    let terminal_safe_fall_length = TERMINAL_SAFE_FALL_VELOCITY_LIKE_CPP
-        * TERMINAL_SAFE_FALL_VELOCITY_LIKE_CPP
-        / (2.0 * GRAVITY_LIKE_CPP);
-    let terminal_length =
-        TERMINAL_VELOCITY_LIKE_CPP * TERMINAL_VELOCITY_LIKE_CPP / (2.0 * GRAVITY_LIKE_CPP);
-    let terminal_safe_fall_time = TERMINAL_SAFE_FALL_VELOCITY_LIKE_CPP / GRAVITY_LIKE_CPP;
-    let terminal_fall_time = TERMINAL_VELOCITY_LIKE_CPP / GRAVITY_LIKE_CPP;
-
-    if is_safe_fall {
-        if path_length >= terminal_safe_fall_length {
-            (path_length - terminal_safe_fall_length) / TERMINAL_SAFE_FALL_VELOCITY_LIKE_CPP
-                + terminal_safe_fall_time
-        } else {
-            (2.0 * path_length / GRAVITY_LIKE_CPP).sqrt()
-        }
-    } else if path_length >= terminal_length {
-        (path_length - terminal_length) / TERMINAL_VELOCITY_LIKE_CPP + terminal_fall_time
-    } else {
-        (2.0 * path_length / GRAVITY_LIKE_CPP).sqrt()
-    }
-}
-
-#[must_use]
-pub fn compute_fall_elevation(t_passed: f32, is_safe_fall: bool, start_velocity: f32) -> f32 {
-    let terminal_velocity = if is_safe_fall {
-        TERMINAL_SAFE_FALL_VELOCITY_LIKE_CPP
-    } else {
-        TERMINAL_VELOCITY_LIKE_CPP
-    };
-    let start_velocity = start_velocity.min(terminal_velocity);
-    let terminal_time = if is_safe_fall {
-        TERMINAL_SAFE_FALL_VELOCITY_LIKE_CPP / GRAVITY_LIKE_CPP
-    } else {
-        TERMINAL_VELOCITY_LIKE_CPP / GRAVITY_LIKE_CPP
-    } - start_velocity / GRAVITY_LIKE_CPP;
-
-    if t_passed > terminal_time {
-        terminal_velocity * (t_passed - terminal_time)
-            + start_velocity * terminal_time
-            + GRAVITY_LIKE_CPP * terminal_time * terminal_time * 0.5
-    } else {
-        t_passed * (start_velocity + t_passed * GRAVITY_LIKE_CPP * 0.5)
-    }
-}
-
-fn init_catmull_storage(path: &[Position], smooth: bool, initial_orientation: f32) -> SplineData {
-    let count = path.len();
-    let mut points = vec![Position::ZERO; count + 2];
-    points[1..=count].copy_from_slice(path);
-    points[0] = offset_initial_virtual_point(path[0], initial_orientation);
-    points[count + 1] = path[count - 1];
-    SplineData {
-        points,
-        lengths: Vec::new(),
-        first: 1,
-        last: i32::try_from(count).expect("spline point count fits i32"),
-        cyclic: false,
-        smooth,
-    }
-}
-
-fn init_cyclic_catmull_storage(
-    path: &[Position],
-    smooth: bool,
-    cyclic_point: usize,
-    initial_orientation: f32,
-) -> SplineData {
-    let count = path.len();
-    let mut points = vec![Position::ZERO; count + 3];
-    points[1..=count].copy_from_slice(path);
-    points[0] = if cyclic_point == 0 {
-        path[count - 1]
-    } else {
-        offset_initial_virtual_point(path[0], initial_orientation)
-    };
-    points[count + 1] = path[cyclic_point];
-    points[count + 2] = path[cyclic_point + 1];
-    SplineData {
-        points,
-        lengths: Vec::new(),
-        first: 1,
-        last: i32::try_from(count).expect("spline point count fits i32") + 1,
-        cyclic: true,
-        smooth,
-    }
-}
-
-fn init_lengths<F>(spline: &mut SplineData, mut next_length: F)
-where
-    F: FnMut(&SplineData, i32) -> i32,
-{
-    spline.lengths.resize(spline.last as usize + 1, 0);
-    let mut prev_length = 0;
-    for index in spline.first..spline.last {
-        let new_length = next_length(spline, index).max(prev_length);
-        spline.lengths[(index + 1) as usize] = new_length;
-        prev_length = new_length;
-    }
-}
-
-fn offset_initial_virtual_point(position: Position, orientation: f32) -> Position {
-    Position::new(
-        position.x - orientation.cos(),
-        position.y - orientation.sin(),
-        position.z,
-        position.orientation,
-    )
-}
-
-fn segment_length(spline: &SplineData, index: i32) -> f32 {
-    if spline.smooth {
-        let mut current = spline.point(index);
-        let mut length = 0.0;
-        for step in 1..=3 {
-            let next = evaluate_catmullrom(spline, index, step as f32 / 3.0);
-            length += distance_3d(current, next);
-            current = next;
-        }
-        length
-    } else {
-        distance_3d(spline.point(index), spline.point(index + 1))
-    }
-}
-
-fn evaluate_linear(spline: &SplineData, index: i32, u: f32) -> Position {
-    let start = spline.point(index);
-    let end = spline.point(index + 1);
-    Position::new(
-        start.x + (end.x - start.x) * u,
-        start.y + (end.y - start.y) * u,
-        start.z + (end.z - start.z) * u,
-        start.orientation,
-    )
-}
-
-fn evaluate_derivative_linear(spline: &SplineData, index: i32) -> Position {
-    let start = spline.point(index);
-    let end = spline.point(index + 1);
-    Position::xyz(end.x - start.x, end.y - start.y, end.z - start.z)
-}
-
-fn evaluate_catmullrom(spline: &SplineData, index: i32, t: f32) -> Position {
-    let p0 = spline.point(index - 1);
-    let p1 = spline.point(index);
-    let p2 = spline.point(index + 1);
-    let p3 = spline.point(index + 2);
-    let t2 = t * t;
-    let t3 = t2 * t;
-    let x = 0.5
-        * ((2.0 * p1.x)
-            + (-p0.x + p2.x) * t
-            + (2.0 * p0.x - 5.0 * p1.x + 4.0 * p2.x - p3.x) * t2
-            + (-p0.x + 3.0 * p1.x - 3.0 * p2.x + p3.x) * t3);
-    let y = 0.5
-        * ((2.0 * p1.y)
-            + (-p0.y + p2.y) * t
-            + (2.0 * p0.y - 5.0 * p1.y + 4.0 * p2.y - p3.y) * t2
-            + (-p0.y + 3.0 * p1.y - 3.0 * p2.y + p3.y) * t3);
-    let z = 0.5
-        * ((2.0 * p1.z)
-            + (-p0.z + p2.z) * t
-            + (2.0 * p0.z - 5.0 * p1.z + 4.0 * p2.z - p3.z) * t2
-            + (-p0.z + 3.0 * p1.z - 3.0 * p2.z + p3.z) * t3);
-    Position::xyz(x, y, z)
-}
-
-fn evaluate_derivative_catmullrom(spline: &SplineData, index: i32, t: f32) -> Position {
-    let p0 = spline.point(index - 1);
-    let p1 = spline.point(index);
-    let p2 = spline.point(index + 1);
-    let p3 = spline.point(index + 2);
-    let t2 = t * t;
-    let x = 0.5
-        * ((-p0.x + p2.x)
-            + 2.0 * (2.0 * p0.x - 5.0 * p1.x + 4.0 * p2.x - p3.x) * t
-            + 3.0 * (-p0.x + 3.0 * p1.x - 3.0 * p2.x + p3.x) * t2);
-    let y = 0.5
-        * ((-p0.y + p2.y)
-            + 2.0 * (2.0 * p0.y - 5.0 * p1.y + 4.0 * p2.y - p3.y) * t
-            + 3.0 * (-p0.y + 3.0 * p1.y - 3.0 * p2.y + p3.y) * t2);
-    let z = 0.5
-        * ((-p0.z + p2.z)
-            + 2.0 * (2.0 * p0.z - 5.0 * p1.z + 4.0 * p2.z - p3.z) * t
-            + 3.0 * (-p0.z + 3.0 * p1.z - 3.0 * p2.z + p3.z) * t2);
-    Position::xyz(x, y, z)
-}
-
-fn wrap_angle_0_2pi(angle: f32) -> f32 {
-    angle.rem_euclid(2.0 * PI)
-}
-
-fn distance_3d(left: Position, right: Position) -> f32 {
-    let dx = left.x - right.x;
-    let dy = left.y - right.y;
-    let dz = left.z - right.z;
-    (dx * dx + dy * dy + dz * dz).sqrt()
-}
-
-fn ms_to_sec(ms: i32) -> f32 {
-    ms as f32 / 1000.0
-}
-
-#[cfg(test)]
-#[path = "spline/tests/mod.rs"]
-mod tests;
